requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.1
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
faiss-cpu==1.7.4
tiktoken==0.5.1
openai==1.3.7
requests==2.31.0 